Create ble-task
